refactor: Update import statements for Sample type
